Re-sync dependency versions with react-redux
