+++ .babelrc
@@ -1,4 +1,6 @@
 {
-  "stage": 0,
-  "loose": "all"
+  "presets": ["es2015-loose", "stage-0", "react"],
+  "plugins": [
+    "transform-decorators-legacy"
+  ]
 }
