+++ .eslintignore
@@ -1,2 +1,3 @@
 lib
 node_modules
+webpack.config.js
