+++ package.json
@@ -8,7 +8,7 @@
     "clean": "rimraf lib dist coverage",
     "lint": "eslint src test",
     "prepublish": "npm run clean && npm run lint && npm run build",
-    "test": "mocha --compilers js:babel/register --recursive --require ./test/setup.js",
+    "test": "mocha --compilers js:babel-core/register --recursive --require ./test/setup.js",
     "test:watch": "npm test -- --watch",
     "test:cov": "babel-node ./node_modules/isparta/bin/isparta cover ./node_modules/mocha/bin/_mocha -- --recursive"
   },
@@ -35,15 +35,19 @@
   },
   "homepage": "https://github.com/heroku/react-refetch",
   "devDependencies": {
-    "babel": "^5.8.21",
-    "babel-core": "^5.8.22",
-    "babel-eslint": "^4.1.7",
-    "babel-loader": "^5.3.2",
-    "eslint": "^1.10.0",
-    "eslint-config-rackt": "^1.1.0",
-    "eslint-plugin-react": "^3.9.0",
+    "babel-cli": "^6.3.17",
+    "babel-core": "^6.3.26",
+    "babel-eslint": "^5.0.0-beta9",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-transform-decorators-legacy": "^1.2.0",
+    "babel-preset-es2015-loose": "^6.1.4",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "eslint": "^1.7.1",
+    "eslint-config-rackt": "1.1.0",
+    "eslint-plugin-react": "^3.6.3",
     "expect": "^1.8.0",
-    "isparta": "^3.0.3",
+    "isparta": "4.0.0",
     "istanbul": "^0.3.17",
     "jsdom": "~5.4.3",
     "mocha": "^2.2.5",
